pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runaway-coach"
version = "0.1.0"
description = "AI-powered running coach using LangChain agentic workflows"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.16.0,<1.0.0",
    "supabase>=2.8.0",
    "httpx>=0.28.0",
    "redis>=5.0.1",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["runaway_coach*"]

[tool.setuptools.package-dir]
"runaway_coach" = "."

[project.scripts]
runaway-coach = "runaway_coach.api.main:main"